+++ internal/procscan/worker_test.go
@@ -10,6 +10,7 @@ import (
 	"time"
 
 	dkproc "github.com/yasyf/daemonkit/proc"
+	"github.com/yasyf/daemonkit/trust"
 	"github.com/yasyf/daemonkit/worker"
 )
 
@@ -34,7 +35,7 @@ func TestWorkerScannerCancellationKillsReapsAndUntracks(t *testing.T) {
 	if err != nil {
 		t.Fatal(err)
 	}
-	claim, err := workers.ClaimRuntime()
+	claim, err := workers.ClaimRuntime(trust.VerifierWorkerBudgets())
 	if err != nil {
 		t.Fatal(err)
 	}